TFG/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
dj-database-url==3.0.0
Django==5.1.3
django-crispy-forms==2.3
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
requests==2.32.3
requests==2.32.3
whitenoise==6.7.0
